requirements.txt
black==24.1.1
python-dotenv==1.0.0
